feat(monorepo): create docs and dev-tools
