+++ .gitignore
@@ -1,4 +1,6 @@
 node_modules
 dist
 .cache
-../../.idea
+.idea
+.temp
+cache
