+++ .npmrc
@@ -1,2 +1,1 @@
 auto-install-peers=true
-shamefully-hoist=true
