+++ packages/core/tsconfig.json
@@ -21,7 +21,7 @@
     "declaration": true,
     "declarationDir": "dist",
     "sourceMap": false,
-    "baseUrl": "../..",
+    "baseUrl": ".",
     "allowJs": true,
     "types": [
       "jest",
@@ -29,9 +29,9 @@
     ]
   },
   "exclude": [
-      "../../node_modules"
+      "node_modules"
   ],
   "include": [
-    "core/src/**/*"
+    "./src/**/*"
   ]
 }
