+++ packages/dev-tools/package.json
@@ -1,0 +1,13 @@
+{
+    "name": "@pusula/module.core-devtools",
+    "version": "1.0.0",
+    "scripts": {
+        "dev": "vite",
+        "build": "vite build"
+    },
+    "devDependencies": {
+        "vite": "^4.0.4",
+        "vue": "^3.2.45",
+        "@vitejs/plugin-vue": "^4.0.0"
+    }
+}
