+++ packages/dev-tools/src/App.vue
@@ -1,0 +1,13 @@
+<template>
+
+</template>
+
+<script>
+export default {
+    name: "App.vue"
+};
+</script>
+
+<style scoped>
+
+</style>
