+++ packages/dev-tools/src/main.ts
@@ -1,0 +1,5 @@
+import { createApp } from "vue";
+import App from "./App.vue";
+const app = createApp(App);
+
+app.mount("#app");
