+++ packages/docs/docs/getting-started.md
@@ -1,0 +1,16 @@
+# Getting Started
+
+Install the package
+
+
+npm:
+
+`npm i @pusula/module.core --save`
+
+yarn:
+
+`yarn add @pusula/module.core`
+
+pnpm:
+
+`pnpm i @pusula/module.core`
