+++ packages/docs/docs/index.md
@@ -1,0 +1,15 @@
+---
+layout: home
+
+hero:
+    name: module.core
+    text: Organize your frontend.
+    tagline: Modular, dependency injection, build for enterprise
+    actions:
+    - theme: brand
+      text: Get Started
+      link: /getting-started
+    - theme: alt
+      text: View on GitHub
+      link: https://github.com/PusulaDev/module.core
+---
