+++ packages/docs/package.json
@@ -1,0 +1,13 @@
+{
+    "name": "@pusula/module.core-docs",
+    "private": true,
+    "scripts": {
+        "dev": "vitepress dev docs",
+        "build": "vitepress build docs",
+        "preview": "vitepress preview docs"
+    },
+    "devDependencies": {
+        "vitepress": "latest",
+        "vue": "^3.2.45"
+    }
+}
