+++ packages/swagger-codegen/package.json
@@ -2,6 +2,13 @@
     "name": "@pusula/swagger-codegen",
     "version": "1.0.0",
     "type": "module",
+    "main": "./dist/index.umd.cjs",
+    "module": "./dist/index.js",
+    "exports": {
+        "require": "./dist/index.umd.cjs",
+        "import": "./dist/index.js"
+    },
+    "types": "./dist/index.d.ts",
     "scripts": {
         "build": "vite build"
     },
