--- packages/swagger-codegen/src/__tests__/generate.spec.ts
@@ -1,7 +1,0 @@
-import { describe, it } from "vitest";
-
-describe("Generate",() => {
-    it("should generate codes with default templates",() => {
-
-    })
-})
