+++ packages/swagger-codegen/tsconfig.json
@@ -7,10 +7,14 @@
         "moduleResolution": "node",
         "skipLibCheck": true,
         "sourceMap": true,
-        "types": [
-            "node"
-        ],
+        "skipDefaultLibCheck": true,
+        "importHelpers": true,
+        "allowSyntheticDefaultImports": true,
+        "alwaysStrict": true
     },
+    "include": [
+        "src/**/*"
+    ],
     "exclude": [
         "node_modules"
     ]
